Update `M20260206Tests` to assert `component_id` column type as `tinyint`

+++ test/LogixDb.Data.SqlServer.Tests/Migrations/M20260206Tests.cs
@@ -99,7 +99,7 @@ public class M20260206Tests : SqlServerTestFixture
         {
             await AssertTableExists("target_component");
 
-            await AssertColumnDefinition("target_component", "component_id", "int");
+            await AssertColumnDefinition("target_component", "component_id", "tinyint");
             await AssertColumnDefinition("target_component", "component_name", "nvarchar");
 
             await AssertPrimaryKey("target_component", "component_id");
